Add Fetch fuction for all Pages
