+++ client/src/Views/UserPage.tsx
@@ -1,46 +1,78 @@
 import React, {  } from 'react';
 import './Pages.css';
-import NavMenuLayout from "../components/NavBar";
+import NavMenu from "../components/NavBar";
 import "bootstrap/dist/css/bootstrap.css";
 import { UserPageDto } from "../Interfaces";
-import NavMenu from "../components/NavBarLogic";
+import { Navigate } from "react-router-dom";
 
 
 class UserPage extends React.Component<{}, UserPageDto> {
 	constructor(props: any) {
 		super(props);
-
+		this.logoutHandler = this.logoutHandler.bind(this);
 		//Initalize state variables
 		this.state = {
 			username: "",
 			picfilepath: "",
 			points: 0,
+			//This can be set to true because it should only be on pages when you are logged in
+			isLoggedIn: true,
 		};
 	}
-	dataLoadIn(){
-		const requestUserData = {
-			method:"GET",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify({
-				username: this.state.username,
-				picfilepath: this.state.picfilepath,
-				points: this.state.points,
-			}),
+	//Navbar Logout-Button's Onclick Function
+	logoutHandler() {
+		const requestOptions = {
+			method: "POST",
+			headers: {
+				"Content-Type": "application/json",
+			},
+			withCredentials: true,
 		};
-		fetch("http://localhost:3001/userpage", requestUserData)
-		.then (async (response) => {
+		fetch("http://localhost:3001/logout", requestOptions)
+			.then(async (response) => {
+				const isJson = response.headers
+					.get("content-type")
+					?.includes("application/json");
+				const data = isJson && (await response.json());
+				if (response.status == 200) {
+					this.setState({
+						isLoggedIn: false,
+					});
+				}
+			})
+
+			.catch((error) => {
+				console.error("There was an error!", error);
+			});
+	}
+	//Loading In The Users Data
+	loadInData(){
+		fetch("http://localhost:3001/userPage")
+		.then(async (response) => {
 			const isJson = response.headers
-			.get("content-type")
-			?.includes("application/json");
+				.get("content-type")
+				?.includes("application/json");
 			const data = isJson && (await response.json());
 			console.log(data);
+			this.setState({username: data.userdata.username});
+			this.setState({picfilepath: data.userdata.picfilepath});
+			this.setState({points: data.userdata.points});
 		})
-
-	};
+		.catch((error) => {
+			console.error("There was an error!", error);
+		});
+	}
+	componentDidMount(){
+		this.loadInData()
+	}
+	//Rendering Page
     render(): React.ReactNode {
+		if (!this.state.isLoggedIn) {
+			return <Navigate to="/login" replace={true} />;
+		} else {
 		return (
 			<div key={"userPage"}>
-				<NavMenu username="" profilePicturePath="" />
+				<NavMenu username={this.state.username} picfilepath={this.state.picfilepath} logoutHandler={this.logoutHandler}/>
 				<div className="text-center mt-3">
 					<div>
 						<h1>{this.state.points} <span id='pont'>Zöldpont</span>-od van</h1>
@@ -63,10 +95,6 @@ class UserPage extends React.Component<{}, UserPageDto> {
 				</div>
 			</div>
 		);
-	}
+	}}
 }
 export default UserPage;
-
-function dataLoadIn() {
-	throw new Error('Function not implemented.');
-}
